fix: change totalCartValue type int with BigDecimal

+++ src/main/java/com/pawelapps/ecommerce/entity/Summary.java
@@ -3,6 +3,8 @@ package com.pawelapps.ecommerce.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "summary")
 @Setter
@@ -17,7 +19,7 @@ public class Summary {
     private Long id;
 
     @Column(name = "total_price")
-    private int totalCartValue;
+    private BigDecimal totalCartValue;
 
     @Column(name = "total_quantity")
     private int totalQuantityOfProducts;
